+++ ui/src/components/dashboard/components/Editor.vue
